+++ apps/docs/src/content/03-components/form-controls/search-field/examples/align.tsx
@@ -1,0 +1,10 @@
+import {
+  Align,
+  Button,
+  SearchField,
+} from "@mittwald/flow-react-components";
+
+<Align>
+  <SearchField />
+  <Button>Hinzufügen</Button>
+</Align>;