+++ packages/components/src/components/FieldError/FieldError.tsx
@@ -1,28 +1,24 @@
-import type { PropsWithChildren } from "react";
-import React from "react";
-import styles from "./FieldError.module.scss";
-import * as Aria from "react-aria-components";
-import clsx from "clsx";
-import { ClearPropsContext } from "@/lib/propsContext";
 import { IconDanger } from "@/components/Icon/components/icons";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
+import ClearPropsContextView from "@/views/ClearPropsContextView";
+import clsx from "clsx";
+import type { PropsWithChildren } from "react";
+import * as Aria from "react-aria-components";
+import styles from "./FieldError.module.scss";
 
 export interface FieldErrorProps
   extends PropsWithChildren<Omit<Aria.FieldErrorProps, "children">>,
     FlowComponentProps {}
 
-/**
- * @flr-generate all
- * @flr-clear-props-context
- */
+/** @flr-generate all */
 export const FieldError = flowComponent("FieldError", (props) => {
   const { children, className, ...rest } = props;
 
   const rootClassName = clsx(styles.fieldError, className);
 
   return (
-    <ClearPropsContext>
+    <ClearPropsContextView>
       <Aria.FieldError {...rest} className={rootClassName}>
         {({ validationErrors }) => (
           <>
@@ -31,7 +27,7 @@ export const FieldError = flowComponent("FieldError", (props) => {
           </>
         )}
       </Aria.FieldError>
-    </ClearPropsContext>
+    </ClearPropsContextView>
   );
 });