+++ packages/components/src/components/Icon/Icon.tsx
@@ -1,31 +1,31 @@
-import type { PropsWithChildren, SVGAttributes } from "react";
-import React from "react";
-import styles from "./Icon.module.scss";
-import clsx from "clsx";
-import { ClearPropsContext } from "@/lib/propsContext";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
 import { cloneElement } from "@/lib/react/cloneElement";
+import type { PropsWithStatus } from "@/lib/types/props";
+import ClearPropsContextView from "@/views/ClearPropsContextView";
+import clsx from "clsx";
+import type { PropsWithChildren, SVGAttributes } from "react";
+import React from "react";
+import styles from "./Icon.module.scss";
 
 type SvgAttributeProps = SVGAttributes<SVGSVGElement>;
 
 export interface IconProps
   extends PropsWithChildren<Omit<SvgAttributeProps, "name">>,
-    FlowComponentProps {
+    FlowComponentProps,
+    PropsWithStatus {
   /** The size of the icon. @default "m" */
   size?: "s" | "m" | "l";
 }
 
-/**
- * @flr-generate all
- * @flr-clear-props-context
- */
+/** @flr-generate all */
 export const Icon = flowComponent("Icon", (props) => {
   const {
     className,
     "aria-label": ariaLabel,
     children,
     size = "m",
+    status,
     ref: ignoredRef,
     ...svgAttributes
   } = props;
@@ -36,7 +36,12 @@ export const Icon = flowComponent("Icon", (props) => {
     role: "img",
     "aria-hidden": !ariaLabel,
     "aria-label": ariaLabel,
-    className: clsx(styles.icon, className, styles[`size-${size}`]),
+    className: clsx(
+      styles.icon,
+      className,
+      styles[`size-${size}`],
+      status && styles[status],
+    ),
   };
 
   const iconElement = React.Children.toArray(children)[0];
@@ -50,9 +55,9 @@ export const Icon = flowComponent("Icon", (props) => {
   }
 
   return (
-    <ClearPropsContext>
+    <ClearPropsContextView>
       {cloneElement(iconElement, iconProps)}
-    </ClearPropsContext>
+    </ClearPropsContextView>
   );
 });