+++ packages/components/src/components/SearchField/SearchField.tsx
@@ -1,27 +1,24 @@
-import type { PropsWithChildren } from "react";
-import React from "react";
-import * as Aria from "react-aria-components";
-import formFieldStyles from "../FormField/FormField.module.scss";
-import styles from "./SearchField.module.scss";
-import clsx from "clsx";
-import type { PropsContext } from "@/lib/propsContext";
-import { ClearPropsContext, PropsContextProvider } from "@/lib/propsContext";
+import { Button } from "@/components/Button";
 import { FieldError } from "@/components/FieldError";
+import { IconClose, IconSearch } from "@/components/Icon/components/icons";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
-import { Button } from "@/components/Button";
-import { IconClose, IconSearch } from "@/components/Icon/components/icons";
-import locales from "./locales/*.locale.json";
+import type { PropsContext } from "@/lib/propsContext";
+import { PropsContextProvider } from "@/lib/propsContext";
+import ClearPropsContextView from "@/views/ClearPropsContextView";
+import clsx from "clsx";
+import type { PropsWithChildren } from "react";
 import { useLocalizedStringFormatter } from "react-aria";
+import * as Aria from "react-aria-components";
+import formFieldStyles from "../FormField/FormField.module.scss";
+import locales from "./locales/*.locale.json";
+import styles from "./SearchField.module.scss";
 
 export interface SearchFieldProps
   extends PropsWithChildren<Omit<Aria.SearchFieldProps, "children">>,
     FlowComponentProps {}
 
-/**
- * @flr-generate all
- * @flr-clear-props-context
- */
+/** @flr-generate all */
 export const SearchField = flowComponent<"SearchField", HTMLInputElement>(
   "SearchField",
   (props) => {
@@ -51,7 +48,7 @@ export const SearchField = flowComponent<"SearchField", HTMLInputElement>(
     };
 
     return (
-      <ClearPropsContext>
+      <ClearPropsContextView>
         <Aria.SearchField
           aria-label={searchText}
           {...rest}
@@ -77,7 +74,7 @@ export const SearchField = flowComponent<"SearchField", HTMLInputElement>(
           </div>
           <FieldError className={formFieldStyles.fieldError} />
         </Aria.SearchField>
-      </ClearPropsContext>
+      </ClearPropsContextView>
     );
   },
 );