+++ packages/remote-react-components/src/auto-generated/CountryOptions.ts
@@ -8,9 +8,6 @@ export { type RemoteCountryOptionsElement } from "@mittwald/flow-remote-elements
 export const CountryOptions = createFlowRemoteComponent(
   "flr-country-options",
   "CountryOptions",
-  {
-    clearPropsContext: false,
-  },
   RemoteCountryOptionsElement,
   {
     slotProps: {