+++ packages/remote-react-components/src/auto-generated/FileCard.ts
@@ -8,9 +8,6 @@ export { type RemoteFileCardElement } from "@mittwald/flow-remote-elements";
 export const FileCard = createFlowRemoteComponent(
   "flr-file-card",
   "FileCard",
-  {
-    clearPropsContext: false,
-  },
   RemoteFileCardElement,
   {
     slotProps: {