+++ packages/remote-react-components/src/auto-generated/LabeledValue.ts
@@ -8,9 +8,6 @@ export { type RemoteLabeledValueElement } from "@mittwald/flow-remote-elements";
 export const LabeledValue = createFlowRemoteComponent(
   "flr-labeled-value",
   "LabeledValue",
-  {
-    clearPropsContext: true,
-  },
   RemoteLabeledValueElement,
   {
     slotProps: {