+++ packages/remote-react-components/src/auto-generated/LightBox.ts
@@ -8,9 +8,6 @@ export { type RemoteLightBoxElement } from "@mittwald/flow-remote-elements";
 export const LightBox = createFlowRemoteComponent(
   "flr-light-box",
   "LightBox",
-  {
-    clearPropsContext: true,
-  },
   RemoteLightBoxElement,
   {
     slotProps: {