+++ packages/remote-react-components/src/auto-generated/Option.ts
@@ -8,9 +8,6 @@ export { type RemoteOptionElement } from "@mittwald/flow-remote-elements";
 export const Option = createFlowRemoteComponent(
   "flr-option",
   "Option",
-  {
-    clearPropsContext: false,
-  },
   RemoteOptionElement,
   {
     slotProps: {