+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.77](https://github.com/mittwald/flow/compare/0.2.0-alpha.76...0.2.0-alpha.77) (2025-03-07)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# [0.2.0-alpha.76](https://github.com/mittwald/flow/compare/0.2.0-alpha.75...0.2.0-alpha.76) (2025-03-06)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# [0.2.0-alpha.75](https://github.com/mittwald/flow/compare/0.2.0-alpha.74...0.2.0-alpha.75) (2025-03-06)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# [0.2.0-alpha.74](https://github.com/mittwald/flow/compare/0.2.0-alpha.73...0.2.0-alpha.74) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# [0.2.0-alpha.73](https://github.com/mittwald/flow/compare/0.2.0-alpha.72...0.2.0-alpha.73) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# [0.2.0-alpha.72](https://github.com/mittwald/flow/compare/0.2.0-alpha.71...0.2.0-alpha.72) (2025-03-05)
+
+### Features
+
+* **Icon:** support status in icon ([#1241](https://github.com/mittwald/flow/issues/1241)) ([6f7028d](https://github.com/mittwald/flow/commit/6f7028d88f2cc29b7b6447bc2bdd38c29868a581))
+
+# [0.2.0-alpha.71](https://github.com/mittwald/flow/compare/0.2.0-alpha.70...0.2.0-alpha.71) (2025-03-05)
+
+### Bug Fixes
+
+* outline style for safari ([#1242](https://github.com/mittwald/flow/issues/1242)) ([d6adf34](https://github.com/mittwald/flow/commit/d6adf34f80e5bbd73d70de1afc5fdd15eb3122a0))
+
+# [0.2.0-alpha.70](https://github.com/mittwald/flow/compare/0.2.0-alpha.69...0.2.0-alpha.70) (2025-03-04)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
 # [0.2.0-alpha.69](https://github.com/mittwald/flow/compare/0.2.0-alpha.68...0.2.0-alpha.69) (2025-03-04)
 
 **Note:** Version bump only for package @mittwald/flow-design-tokens