+++ packages/remote-react-components/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.77](https://github.com/mittwald/flow/compare/0.2.0-alpha.76...0.2.0-alpha.77) (2025-03-07)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.76](https://github.com/mittwald/flow/compare/0.2.0-alpha.75...0.2.0-alpha.76) (2025-03-06)
+
+### Bug Fixes
+
+* fix props context not properly cleared in remote components ([#1240](https://github.com/mittwald/flow/issues/1240)) ([4152991](https://github.com/mittwald/flow/commit/4152991202061163e4cbf1e66b85bb8309abfc4f))
+
+# [0.2.0-alpha.75](https://github.com/mittwald/flow/compare/0.2.0-alpha.74...0.2.0-alpha.75) (2025-03-06)
+
+### Bug Fixes
+
+* use patched remote packages ([#1246](https://github.com/mittwald/flow/issues/1246)) ([09fa922](https://github.com/mittwald/flow/commit/09fa922626fb0f799985669d1118792f9aa7ebba))
+
+# [0.2.0-alpha.74](https://github.com/mittwald/flow/compare/0.2.0-alpha.73...0.2.0-alpha.74) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.73](https://github.com/mittwald/flow/compare/0.2.0-alpha.72...0.2.0-alpha.73) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.72](https://github.com/mittwald/flow/compare/0.2.0-alpha.71...0.2.0-alpha.72) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.71](https://github.com/mittwald/flow/compare/0.2.0-alpha.70...0.2.0-alpha.71) (2025-03-05)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.70](https://github.com/mittwald/flow/compare/0.2.0-alpha.69...0.2.0-alpha.70) (2025-03-04)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
 # [0.2.0-alpha.69](https://github.com/mittwald/flow/compare/0.2.0-alpha.68...0.2.0-alpha.69) (2025-03-04)
 
 **Note:** Version bump only for package @mittwald/flow-remote-react-components